Fix bug where Cache.fetch returned random objects
The Weakref implementation of CacheEntry works well, except when Ruby
allocates a new Object with an object_id whose CacheEntry-finalizer has
not yet run. Therefore we need to double-check whether the Object
returned by ObjectSpace._id2ref really is the one we're looking for.
This is done using Object#hash. And since that value may change if an
object is modified, we need to update CacheEntry whenever a modified
object is stored.